+++ plugins/pelm-linux.el
@@ -17,11 +17,8 @@
 
 (when (string-match "linux" system-configuration)
   (define-key pelm-keymap "uf" 'pelm-linux-toggle-fullscreen)
-  (set-face-font 'default  "Ubuntu Mono-13")
+  (set-face-font 'default "WenQuanYi Zen Hei Mono-12")
   )
 
-;这里补充几点：
-;首先你需要安装文泉驿正黑（也可以配置成你喜欢的其它字体，例如微软雅黑或幼圆）
-;我在这里设置环境为 utf8 ，其实是个人的一些小偏执，如果你不设置，默认对中文Windows支持的很好，但是我需要在各种操作系统间切换，保持我所有的文本都是UTF8编码对我非常重要，所以这里就强制了。
 (provide 'pelm-linux)
 ;;; ends pelm-linux.el here