+++ plugins/pelm-lisp.el
@@ -9,19 +9,16 @@
 ;; This file is not part of GNU Emacs
 
 ;;; Code:
+;; configuration
 
 
-(add-hook 'emacs-lisp-mode-hook 'highlight-parentheses-mode )
-;(add-hook 'clojure-mode-hook 'autopair-mode)
+(add-hook 'emacs-lisp-mode-hook 'subword-mode)
+(add-hook 'emacs-lisp-mode-hook 'paredit-mode)
 
-(add-hook 'slime-repl-mode-hook
-          (defun clojure-mode-slime-font-lock ()
-            (require 'clojure-mode)
-            (let (font-lock-mode)
-              (clojure-mode-font-lock-setup))))
+(add-hook 'emacs-lisp-mode-hook 'rainbow-delimiters-mode)
 
-(setq swank-clojure-classpath (directory-files "~/.emacs.d/vendor/jars" t ".jar$"))
+
+
 (provide 'pelm-lisp)
-
 ;;; pelm-lisp.el ends here